complete TRT-8.4

+++ cookbook/08-Tool/trex/trex/df_preprocessing.py
@@ -22,139 +22,139 @@ import pandas as pd
 from .activations import *
 
 layer_attributes = {
-    'Convolution': {
-        'subtype': 'subtype',
-        'Groups': 'attr.groups',
-        'OutMaps': 'attr.out_maps',
-        'Stride': 'attr.stride',
-        'Kernel': 'attr.kernel',
-        'HasBias': 'attr.has_bias',
-        'Activation': 'attr.activation',
-        'HasReLU': 'attr.has_relu',
-        'PaddingMode': None,  # 'attr.padding_mode',
-        'PrePadding': None,  # 'attr.pre_padding',
-        'PostPadding': None,  # 'attr.post_padding',
-        'Dilation': 'attr.dilation',
-        'AllowSparse': None,
-        'Weights': 'Weights',
-        'Bias': None,
-        'Dimensions': None,
-        'ConvolutionTacticIndex': None,
-        'Operations': 'attr.operations',
-        'NbOperations': 'attr.n_operations',
+    "Convolution": {
+        "subtype": "subtype",
+        "Groups": "attr.groups",
+        "OutMaps": "attr.out_maps",
+        "Stride": "attr.stride",
+        "Kernel": "attr.kernel",
+        "HasBias": "attr.has_bias",
+        "Activation": "attr.activation",
+        "HasReLU": "attr.has_relu",
+        "PaddingMode": None,  # "attr.padding_mode",
+        "PrePadding": None,  # "attr.pre_padding",
+        "PostPadding": None,  # "attr.post_padding",
+        "Dilation": "attr.dilation",
+        "AllowSparse": None,
+        "Weights": "Weights",
+        "Bias": None,
+        "Dimensions": None,
+        "ConvolutionTacticIndex": None,
+        "Operations": "attr.operations",
+        "NbOperations": "attr.n_operations",
     },
-    'PointWiseV2': {
-        'Operations': 'attr.operations',
-        'NbOperations': 'attr.n_operations',
-        'NbInputArgs': 'attr.n_input_args',
-        'InputArgs': None,
-        'NbOutputVars': None,
-        'OutputVars': None,
-        'NbLiterals': None,
-        'Literals': None,
-        'NbParams': None,
-        'Params': None,
+    "PointWiseV2": {
+        "Operations": "attr.operations",
+        "NbOperations": "attr.n_operations",
+        "NbInputArgs": "attr.n_input_args",
+        "InputArgs": None,
+        "NbOutputVars": None,
+        "OutputVars": None,
+        "NbLiterals": None,
+        "Literals": None,
+        "NbParams": None,
+        "Params": None,
     },
-    'PointWise': {
-        'Instructions': 'attr.operations',
-        'NbInstructions': 'attr.n_operations',
-        'NbInputs': 'attr.n_input_args',
-        'InputArgs': None,
-        'NbOutputVars': None,
-        'OutputVars': None,
-        'NbLiterals': None,
-        'Literals': None,
-        'NbParams': None,
-        'Params': None,
+    "PointWise": {
+        "Instructions": "attr.operations",
+        "NbInstructions": "attr.n_operations",
+        "NbInputs": "attr.n_input_args",
+        "InputArgs": None,
+        "NbOutputVars": None,
+        "OutputVars": None,
+        "NbLiterals": None,
+        "Literals": None,
+        "NbParams": None,
+        "Params": None,
     },
-    'Reformat': {
-        'Origin': 'attr.origin',
+    "Reformat": {
+        "Origin": "attr.origin",
     },
-    'Pooling': {
-        'PoolingType': 'attr.pooling_type',
-        'WindowSize': 'attr.window_size',
-        'AverageCountExcludesPadding': None,
-        'PaddingMode': 'attr.padding_mode',
-        'Stride': 'attr.stride',
-        'PrePadding': None,
-        'PostPadding': None,
-        'BlendFactor': None,
+    "Pooling": {
+        "PoolingType": "attr.pooling_type",
+        "WindowSize": "attr.window_size",
+        "AverageCountExcludesPadding": None,
+        "PaddingMode": "attr.padding_mode",
+        "Stride": "attr.stride",
+        "PrePadding": None,
+        "PostPadding": None,
+        "BlendFactor": None,
     },
-    'Scale': {
-        'Mode': 'attr.mode',
-        'Shift': 'attr.shift',
-        'Scale': 'attr.scale',
-        'Power': 'attr.power',
-        'Activation': 'attr.activation',
-        'ChannelAxis': 'attr.ch_axis',
+    "Scale": {
+        "Mode": "attr.mode",
+        "Shift": "attr.shift",
+        "Scale": "attr.scale",
+        "Power": "attr.power",
+        "Activation": "attr.activation",
+        "ChannelAxis": "attr.ch_axis",
     },
-    'Shuffle': {
-        'FirstTranspose': 'attr.first_transpose',
-        'SecondTranspose': 'attr.second_transpose',
-        'Reshape': 'attr.reshape',
-        'ZeroIsPlaceholder': None,
-        'ParameterSubType': None,
+    "Shuffle": {
+        "FirstTranspose": "attr.first_transpose",
+        "SecondTranspose": "attr.second_transpose",
+        "Reshape": "attr.reshape",
+        "ZeroIsPlaceholder": None,
+        "ParameterSubType": None,
     },
-    'Resize': {
-        'ResizeMode': 'attr.mode',
-        'ResizeScales': 'attr.scales',
-        'NNRounding': 'attr.scale',
-        'Start': None,
-        'CoordTransform': None,
-        'ResizeSelector': None,
+    "Resize": {
+        "ResizeMode": "attr.mode",
+        "ResizeScales": "attr.scales",
+        "NNRounding": "attr.scale",
+        "Start": None,
+        "CoordTransform": None,
+        "ResizeSelector": None,
     },
-    'Slice': {
-        'Start': 'attr.start',
-        'Stride': 'attr.stride',
-        'Size': 'attr.size',
-        'Mode': 'attr.mode',
-        'negativeInfinityPadding': None,
+    "Slice": {
+        "Start": "attr.start",
+        "Stride": "attr.stride",
+        "Size": "attr.size",
+        "Mode": "attr.mode",
+        "negativeInfinityPadding": None,
     },
-    'Common': {
-        'ParameterType': None,
-        'weights': None,
-        'dimensions': None,
-        'TacticValue': None,
+    "Common": {
+        "ParameterType": None,
+        "weights": None,
+        "dimensions": None,
+        "TacticValue": None,
     },
 }
 
 def __fix_type(df: pd.DataFrame):
-    df.rename(columns={'LayerType': 'subtype'}, inplace=True)
+    df.rename(columns={"LayerType": "subtype"}, inplace=True)
     try:
-        df['type'] = df.ParameterType.fillna(value=df.subtype)
-        df.drop(['ParameterType'], axis=1, inplace=True)
+        df["type"] = df.ParameterType.fillna(value=df.subtype)
+        df.drop(["ParameterType"], axis=1, inplace=True)
     except AttributeError:
         pass
 
 def __fix_tactic(df: pd.DataFrame):
-    df.rename(columns={'TacticName': 'tactic'}, inplace=True)
+    df.rename(columns={"TacticName": "tactic"}, inplace=True)
     try:
-        df['tactic'] = df.tactic.fillna(value='TensorRT')
+        df["tactic"] = df.tactic.fillna(value="TensorRT")
     except AttributeError:
-        df['tactic'] = 'TensorRT'
+        df["tactic"] = "TensorRT"
 
 def __fix_columns_types(df: pd.DataFrame):
     int_cols = [
-        'Groups',
-        'OutMaps',
-        'HasBias',
-        'HasReLU',
-        'AllowSparse',
-        'NbInputArgs',
-        'NbOutputVars',
-        'NbParams',
-        'NbLiterals',
+        "Groups",
+        "OutMaps",
+        "HasBias",
+        "HasReLU",
+        "AllowSparse",
+        "NbInputArgs",
+        "NbOutputVars",
+        "NbParams",
+        "NbLiterals",
     ]
     for col in int_cols:
         try:
             df[col] = df[col].fillna(value=0)
-            df[col] = df[col].astype('int32')
+            df[col] = df[col].astype("int32")
         except KeyError:
             pass
     df.fillna("", inplace=True)
 
 def __fix_output_precision(df: pd.DataFrame):
-    df['output_precision'] = [Activation(outputs[0]).precision for outputs in df['Outputs']]
+    df["output_precision"] = [Activation(outputs[0]).precision for outputs in df["Outputs"]]
 
 def fix_df(df: pd.DataFrame):
     """One-time preprocessing of the DF.
@@ -172,17 +172,17 @@ def clean_io(df: pd.DataFrame):
         inputs, outputs = create_activations(layer)
         if len(inputs) > 0:
             inp_str = ", ".join([inp.format for inp in inputs])
-            df.loc[index, 'Inputs'] = inp_str
-        df.loc[index, 'Outputs'] = outputs[0].format
+            df.loc[index, "Inputs"] = inp_str
+        df.loc[index, "Outputs"] = outputs[0].format
 
 def filter_by_layer(df: pd.DataFrame, layer_type: str):
-    copy_cols = ['Name', 'type', 'precision', 'tactic', 'latency.pct_time', 'latency.avg_time', 'total_io_size_bytes', 'total_footprint_bytes', 'Inputs', 'Outputs', 'subtype']
+    copy_cols = ["Name", "type", "precision", "tactic", "latency.pct_time", "latency.avg_time", "total_io_size_bytes", "total_footprint_bytes", "Inputs", "Outputs", "subtype"]
     try:
         attrs = layer_attributes[layer_type]
         copy_cols += [k for k, v in attrs.items() if v is not None]
         # Pointwise and Pointwise V2 layers have the same layer type.
-        if layer_type == 'PointWise':
-            attrs = layer_attributes['PointWiseV2']
+        if layer_type == "PointWise":
+            attrs = layer_attributes["PointWiseV2"]
             copy_cols += [k for k, v in attrs.items() if v is not None]
     except KeyError:
         pass
@@ -194,16 +194,16 @@ def filter_by_layer(df: pd.DataFrame, layer_type: str):
     copy_cols = set(copy_cols) & set(layers.columns)
     layers = layers[copy_cols]
 
-    if layer_type == 'Convolution':
+    if layer_type == "Convolution":
         layers.rename(columns=layer_attributes[layer_type], inplace=True)
-        layers['attr.kernel'] = tuple(layers['attr.kernel'])
+        layers["attr.kernel"] = tuple(layers["attr.kernel"])
         annotate_convolutions(layers)
-    if layer_type == 'PointWise':
+    if layer_type == "PointWise":
         # The original JSON file handle PointWise and PointWise V2 as two subtypes.
-        pw = layers[layers['subtype'] == 'PointWise'].copy()
-        pw.rename(columns=layer_attributes['PointWise'], inplace=True)
-        pw_v2 = layers[layers['subtype'] == 'PointWiseV2'].copy()
-        pw_v2.rename(columns=layer_attributes['PointWiseV2'], inplace=True)
+        pw = layers[layers["subtype"] == "PointWise"].copy()
+        pw.rename(columns=layer_attributes["PointWise"], inplace=True)
+        pw_v2 = layers[layers["subtype"] == "PointWiseV2"].copy()
+        pw_v2.rename(columns=layer_attributes["PointWiseV2"], inplace=True)
         layers = pd.concat((pw, pw_v2))
     else:
         try:
@@ -215,7 +215,7 @@ def filter_by_layer(df: pd.DataFrame, layer_type: str):
 def change_col_order(df: pd.DataFrame):
     """Change the dataframe columns-order (place common fields earlier)"""
     cols = df.columns.to_list()
-    common_cols = list(('Name', 'type', 'Inputs', 'Outputs', 'latency.avg_time', 'latency.pct_time', 'total_footprint_bytes', 'tactic'))
+    common_cols = list(("Name", "type", "Inputs", "Outputs", "latency.avg_time", "latency.pct_time", "total_footprint_bytes", "tactic"))
     common_cols = [col for col in common_cols if col in cols]
     cols = common_cols + [col for col in cols if col not in common_cols]
     df = df[cols]
@@ -239,7 +239,7 @@ def clean_for_display(df: pd.DataFrame):
     """Prepare the dataframe for display"""
     df = clean_df(df.copy(), inplace=True)
     df = change_col_order(df)
-    drop_columns(df, columns=['subtype', 'TacticValue', 'precision', 'total_io_size_bytes'])
+    drop_columns(df, columns=["subtype", "TacticValue", "precision", "total_io_size_bytes"])
     return df
 
 def annotate_convolutions(convs: pd.DataFrame):
@@ -250,8 +250,8 @@ def annotate_convolutions(convs: pd.DataFrame):
         N, C, H, W = inputs[0].shape
         # K: number of channels; P: Height; Q: Width
         _, K, P, Q = outputs[0].shape
-        R, S = convs.loc[index, 'attr.kernel']
-        G = convs.loc[index, 'attr.groups']
+        R, S = convs.loc[index, "attr.kernel"]
+        G = convs.loc[index, "attr.groups"]
         weights_vol = (K * C * R * S) / G
         input_vol = N * C * H * W
         output_vol = N * K * P * Q
@@ -260,21 +260,21 @@ def annotate_convolutions(convs: pd.DataFrame):
         weights_bytes = weights_vol * inputs[0].data_size
         nb_bytes = input_bytes + weights_bytes + output_bytes
         nb_macs = N * K * P * Q * C * R * S / G
-        convs.loc[index, 'attr.macs'] = nb_macs
+        convs.loc[index, "attr.macs"] = nb_macs
         # Arithmetic intensity: ops/bytes
-        convs.loc[index, 'attr.arithmetic_intensity'] = nb_macs / nb_bytes
-        latency = convs.loc[index, 'latency.avg_time']
-        convs.loc[index, 'attr.compute_efficiency'] = nb_macs / latency
-        convs.loc[index, 'attr.memory_efficiency'] = nb_bytes / latency
+        convs.loc[index, "attr.arithmetic_intensity"] = nb_macs / nb_bytes
+        latency = convs.loc[index, "latency.avg_time"]
+        convs.loc[index, "attr.compute_efficiency"] = nb_macs / latency
+        convs.loc[index, "attr.memory_efficiency"] = nb_bytes / latency
         # Conversion to matrices (M, K) * (K, N)
         M = N * P * Q
         N = K
         K = C * R * S
-        convs.loc[index, 'attr.M'] = M
-        convs.loc[index, 'attr.N'] = N
-        convs.loc[index, 'attr.K'] = K
+        convs.loc[index, "attr.M"] = M
+        convs.loc[index, "attr.N"] = N
+        convs.loc[index, "attr.K"] = K
 
-    convs['attr.macs'] = convs['attr.macs'].astype('int64')
-    convs['attr.M'] = convs['attr.M'].astype('int64')
-    convs['attr.N'] = convs['attr.N'].astype('int64')
-    convs['attr.K'] = convs['attr.K'].astype('int64')
+    convs["attr.macs"] = convs["attr.macs"].astype("int64")
+    convs["attr.M"] = convs["attr.M"].astype("int64")
+    convs["attr.N"] = convs["attr.N"].astype("int64")
+    convs["attr.K"] = convs["attr.K"].astype("int64")
